+++ pkg/apis/options/upstreams.go
@@ -29,8 +29,14 @@ const (
 	// DefaultUpstreamPassHostHeader determines if upstreams will pass the host header by default.
 	DefaultUpstreamPassHostHeader bool = true
 
+	// Static upstreams cannot pass the host header.
+	DefaultStaticPassHostHeader bool = false
+
 	// DefaultUpstreamProxyWebSockets determines if upstreams will proxy websockets by default.
 	DefaultUpstreamProxyWebSockets bool = true
+
+	// Static upstreams cannot proxy websockets.
+	DefaultStaticProxyWebSockets bool = false
 
 	// DefaultUpstreamDisableKeepAlives determines if upstreams will disable keep-alives by default.
 	DefaultUpstreamDisableKeepAlives bool = false
@@ -143,6 +149,12 @@ func (u *Upstream) EnsureDefaults() {
 	if u.Static == nil {
 		u.Static = ptr.To(DefaultUpstreamStatic)
 	}
+	if u.PassHostHeader == nil {
+		u.PassHostHeader = ptr.To(DefaultUpstreamPassHostHeader)
+	}
+	if u.ProxyWebSockets == nil {
+		u.ProxyWebSockets = ptr.To(DefaultUpstreamProxyWebSockets)
+	}
 	if u.FlushInterval == nil {
 		u.FlushInterval = ptr.To(DefaultUpstreamFlushInterval)
 	}
@@ -153,14 +165,15 @@ func (u *Upstream) EnsureDefaults() {
 		u.DisableKeepAlives = ptr.To(DefaultUpstreamDisableKeepAlives)
 	}
 
-	// PassHostHeader and ProxyWebSockets must remain nil for static upstreams
-	// as they don't apply and validation will flag them if set
-	if !ptr.Deref(u.Static, DefaultUpstreamStatic) {
-		if u.PassHostHeader == nil {
-			u.PassHostHeader = ptr.To(DefaultUpstreamPassHostHeader)
-		}
-		if u.ProxyWebSockets == nil {
-			u.ProxyWebSockets = ptr.To(DefaultUpstreamProxyWebSockets)
-		}
+	// Force defaults compatible with static upstreams.
+	// This overrides any user provided values to ensure static upstreams behave correctly.
+	if ptr.Deref(u.Static, DefaultUpstreamStatic) {
+		u.URI = ""
+		u.InsecureSkipTLSVerify = ptr.To(DefaultUpsteamInsecureSkipTLSVerify)
+		u.DisableKeepAlives = ptr.To(DefaultUpstreamDisableKeepAlives)
+		u.PassHostHeader = ptr.To(DefaultStaticPassHostHeader)
+		u.ProxyWebSockets = ptr.To(DefaultStaticProxyWebSockets)
+		u.FlushInterval = ptr.To(DefaultUpstreamFlushInterval)
+		u.Timeout = ptr.To(DefaultUpstreamTimeout)
 	}
 }